+++ .gitignore
@@ -1,6 +1,7 @@
 # custom
 dist
 examples/**/package-lock.json
+*.tgz
 
 # os
 .DS_Store
