+++ package.json
@@ -2,6 +2,7 @@
 	"name": "@prismicio/react",
 	"version": "2.9.1",
 	"description": "React components and hooks to fetch and present Prismic content",
+	"type": "module",
 	"keywords": [
 		"typescript",
 		"prismic",
@@ -11,40 +12,25 @@
 	],
 	"repository": {
 		"type": "git",
-		"url": "ssh://git@github.com/prismicio/prismic-react.git"
+		"url": "git+https://github.com/prismicio/prismic-react.git"
 	},
 	"license": "Apache-2.0",
 	"author": "Prismic <contact@prismic.io> (https://prismic.io)",
 	"sideEffects": false,
 	"exports": {
 		".": {
-			"react-server": {
-				"require": {
-					"types": "./dist/react-server/index.d.ts",
-					"default": "./dist/react-server.cjs"
-				},
-				"import": {
-					"types": "./dist/react-server/index.d.ts",
-					"default": "./dist/react-server.js"
-				}
-			},
-			"default": {
-				"require": {
-					"types": "./dist/index.d.ts",
-					"default": "./dist/index.cjs"
-				},
-				"import": {
-					"types": "./dist/index.d.ts",
-					"default": "./dist/index.js"
-				}
-			}
+			"types": "./dist/index.d.ts",
+			"default": "./dist/index.js"
 		},
 		"./package.json": "./package.json"
 	},
-	"main": "./dist/index.cjs",
-	"module": "./dist/index.js",
-	"react-native": "./dist/index.js",
-	"types": "./dist/index.d.ts",
+	"typesVersions": {
+		"*": {
+			"*": [
+				"./dist/index.d.ts"
+			]
+		}
+	},
 	"files": [
 		"dist",
 		"src"
@@ -65,51 +51,53 @@
 		"size": "size-limit",
 		"test": "npm run lint && npm run types && npm run unit && npm run build && npm run size"
 	},
-	"dependencies": {
-		"@prismicio/richtext": "^2.1.5"
-	},
 	"devDependencies": {
-		"@prismicio/client": "^7.11.0",
+		"@prismicio/client": "^7.12.0",
 		"@prismicio/mock": "0.4.0",
-		"@size-limit/preset-small-lib": "^8.2.4",
-		"@testing-library/react": "^14.0.0",
+		"@rollup/plugin-typescript": "^12.1.1",
+		"@size-limit/preset-small-lib": "^11.1.6",
+		"@testing-library/react": "^16.0.1",
 		"@types/node-fetch": "^3.0.3",
-		"@types/react": "^18.2.9",
-		"@types/react-test-renderer": "^18.0.0",
-		"@typescript-eslint/eslint-plugin": "^5.59.9",
-		"@typescript-eslint/parser": "^5.59.9",
-		"@vitejs/plugin-react": "^4.0.0",
-		"@vitest/coverage-v8": "^0.32.0",
-		"eslint": "^8.42.0",
-		"eslint-config-prettier": "^8.8.0",
-		"eslint-plugin-prettier": "^4.2.1",
-		"eslint-plugin-react": "^7.32.2",
-		"eslint-plugin-react-hooks": "^4.6.0",
-		"eslint-plugin-tsdoc": "^0.2.17",
-		"happy-dom": "^9.20.3",
-		"msw": "^1.2.1",
-		"node-fetch": "^3.3.1",
-		"prettier": "^2.8.8",
-		"prettier-plugin-jsdoc": "^0.4.2",
+		"@types/react": "^18.3.12",
+		"@types/react-test-renderer": "^18.3.0",
+		"@typescript-eslint/eslint-plugin": "^8.14.0",
+		"@typescript-eslint/parser": "^8.14.0",
+		"@vitejs/plugin-react": "^4.3.3",
+		"@vitest/coverage-v8": "^2.1.5",
+		"eslint": "^9.15.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-prettier": "^5.2.1",
+		"eslint-plugin-react": "^7.37.2",
+		"eslint-plugin-react-hooks": "^5.0.0",
+		"eslint-plugin-tsdoc": "^0.3.0",
+		"happy-dom": "^15.11.6",
+		"msw": "^2.6.5",
+		"node-fetch": "^3.3.2",
+		"prettier": "^3.3.3",
+		"prettier-plugin-jsdoc": "^1.3.0",
 		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-test-renderer": "^18.2.0",
-		"rollup-plugin-preserve-directives": "^0.2.0",
-		"size-limit": "^8.2.4",
+		"react-dom": "^18.3.1",
+		"react-test-renderer": "^18.3.1",
+		"rollup-plugin-preserve-directives": "^0.4.0",
+		"rollup-preserve-directives": "^1.1.2",
+		"size-limit": "^11.1.6",
 		"standard-version": "^9.5.0",
-		"typescript": "^5.0.4",
-		"vite": "^4.3.9",
-		"vite-plugin-sdk": "^0.1.1",
-		"vitest": "^0.32.0"
+		"typescript": "^5.6.3",
+		"vite": "^5.4.11",
+		"vite-plugin-sdk": "^0.1.2",
+		"vitest": "^2.1.5"
 	},
 	"peerDependencies": {
-		"@prismicio/client": "^6 || ^7",
+		"@prismicio/client": "^7",
 		"react": "^18 || ^19.0.0-rc.0"
 	},
 	"engines": {
-		"node": ">=12.7.0"
+		"node": ">=18"
 	},
 	"publishConfig": {
 		"access": "public"
+	},
+	"dependencies": {
+		"esm-env": "^1.1.4"
 	}
 }
