+++ vite.config.ts
@@ -1,24 +1,30 @@
 import { defineConfig } from "vite";
-import sdk from "vite-plugin-sdk";
 import react from "@vitejs/plugin-react";
-import preserveDirectives from "rollup-plugin-preserve-directives";
+import typescript from "@rollup/plugin-typescript";
+import preserveDirectives from "rollup-preserve-directives";
+
+import { dependencies, peerDependencies } from "./package.json";
 
 export default defineConfig({
-	plugins: [
-		sdk({
-			internalDependencies: ["@prismicio/client"],
-		}),
-		react(),
-	],
+	plugins: [react()],
 	build: {
 		lib: {
 			entry: {
-				index: "./src/index.ts",
-				"react-server": "./src/react-server/index.ts",
-			},
+				index: "./src/index.ts" },
+			formats: ["es"],
 		},
+		minify: false,
+		sourcemap: true,
 		rollupOptions: {
-			plugins: [preserveDirectives()],
+			output: {
+				preserveModules: true,
+				preserveModulesRoot: "./src",
+			},
+			external: [
+				...Object.keys(dependencies),
+				...Object.keys(peerDependencies),
+			].map((name) => new RegExp(`^${name}(?:\/.*)?$`)),
+			plugins: [typescript(), preserveDirectives()],
 		},
 	},
 	test: {
